Move default implementation to TypeDBStub since they're used by both CoreStub and ClusterServerStub
