+++ connection/core/CoreStub.java
@@ -22,12 +22,8 @@
 package com.vaticle.typedb.client.connection.core;
 
 import com.vaticle.typedb.client.common.rpc.TypeDBStub;
-import com.vaticle.typedb.protocol.CoreDatabaseProto;
-import com.vaticle.typedb.protocol.SessionProto;
-import com.vaticle.typedb.protocol.TransactionProto;
 import com.vaticle.typedb.protocol.TypeDBGrpc;
 import io.grpc.ManagedChannel;
-import io.grpc.stub.StreamObserver;
 
 public class CoreStub extends TypeDBStub {
 
@@ -44,51 +40,6 @@ public class CoreStub extends TypeDBStub {
 
     public static CoreStub create(ManagedChannel channel) {
         return new CoreStub(channel);
-    }
-
-    @Override
-    public CoreDatabaseProto.CoreDatabaseManager.Contains.Res databasesContains(CoreDatabaseProto.CoreDatabaseManager.Contains.Req request) {
-        return resilientCall(() -> blockingStub().databasesContains(request));
-    }
-
-    @Override
-    public CoreDatabaseProto.CoreDatabaseManager.Create.Res databasesCreate(CoreDatabaseProto.CoreDatabaseManager.Create.Req request) {
-        return resilientCall(() -> blockingStub().databasesCreate(request));
-    }
-
-    @Override
-    public CoreDatabaseProto.CoreDatabaseManager.All.Res databasesAll(CoreDatabaseProto.CoreDatabaseManager.All.Req request) {
-        return resilientCall(() -> blockingStub().databasesAll(request));
-    }
-
-    @Override
-    public CoreDatabaseProto.CoreDatabase.Schema.Res databaseSchema(CoreDatabaseProto.CoreDatabase.Schema.Req request) {
-        return resilientCall(() -> blockingStub().databaseSchema(request));
-    }
-
-    @Override
-    public CoreDatabaseProto.CoreDatabase.Delete.Res databaseDelete(CoreDatabaseProto.CoreDatabase.Delete.Req request) {
-        return resilientCall(() -> blockingStub().databaseDelete(request));
-    }
-
-    @Override
-    public SessionProto.Session.Open.Res sessionOpen(SessionProto.Session.Open.Req request) {
-        return resilientCall(() -> blockingStub().sessionOpen(request));
-    }
-
-    @Override
-    public SessionProto.Session.Close.Res sessionClose(SessionProto.Session.Close.Req request) {
-        return resilientCall(() -> blockingStub().sessionClose(request));
-    }
-
-    @Override
-    public SessionProto.Session.Pulse.Res sessionPulse(SessionProto.Session.Pulse.Req request) {
-        return resilientCall(() -> blockingStub().sessionPulse(request));
-    }
-
-    @Override
-    public StreamObserver<TransactionProto.Transaction.Client> transaction(StreamObserver<TransactionProto.Transaction.Server> responseObserver) {
-        return resilientCall(() -> asyncStub().transaction(responseObserver));
     }
 
     @Override
